fix(tls): second-pass review — real diagnostics on Ok(Invalid), refreshed authz state, teardown serialisation
- poll_ready yields Ok(Invalid) for a failed validation (it only Errs on
  timeout/transport or a top-level order error, which Let's Encrypt does not
  set for failed authorizations) — the diagnostics branch never fired for
  the exact case it was built for, and the code fell through to finalize an
  invalid order, reproducing the original opaque failure. Both non-Ready
  outcomes now run the diagnostics and fail with a named status.
- The authorization iterator serves the cached pre-validation snapshot
  (every challenge Pending, no error); each authorization is refresh()ed
  before reading challenge Problems — without it the CA's rejection reason
  was never logged.
- teardown() takes the issuance lock: it could interleave with an in-flight
  order and be resurrected seconds later by the completing issuance.
- Suspension clears retry backoff state (a stale rate-limit floor no longer
  leaks into a post-resume nudge); nudge/rate-limit logs use named
  interpolation per .agents/logging.md; runner tests moved to their own
  tls/tests/runner.rs per the strict layout; module and method docs updated
  to the deadline semantics.
